requirements.txt
boto3==1.33.13
botocore==1.33.13
jmespath==1.0.1
Pillow==9.5.0
python-dateutil==2.9.0.post0
s3transfer==0.8.2
six==1.16.0
urllib3==1.26.20
